Fix Task 0x0A.[ADVANCED].5
WHAT:
Changed function countChange() to a while loop in main function.

WHY:
Recursive function countChange() throws segmentation fault error when high input (e.g. 2147483647) is passed to it.
This is due to a native limitation in C that lets a recursive function call itself only up to a maximum number of times. Each new call duplicates the function in the stack, thus overloading it (filling up available program memory/RAM).
Loops (while, for, etc.) don't have the same issue as the same/existing variable(s) is(are) used, so nothing is duplicated, but rather replaced in each iteration as the need arises.