Do Task 0x0C.[ADVANCED].5
The main() function had sixty (60) lines, but Betty (ALX guidelines) has a restriction of forty (40) lines per function, so it had to be split to form two functions. The second file here (101-mul_longMAIN.c) is the original, with the long main() function, while the other (101-mul.c) has the split main() function for compliance, and is the version the code checker checked.